requirements/core.txt
django==4.2.7  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
drf-spectacular==0.27.1 # https://github.com/tfranzel/drf-spectacular
django-cors-headers==4.3.1 # https://github.com/adamchainz/django-cors-headers
djangorestframework-api-key==3.0.0
psycopg2-binary==2.9.9  # https://github.com/psycopg/psycopg2

celery==5.3.6
redis==5.0.2

python-binance==1.0.19 # https://github.com/sammchardy/python-binance
pydantic==2.6.3 # https://github.com/pydantic/pydantic